+++ hpc/nvidia/cuda/intrinsics/bfloat16.cu
@@ -14,20 +14,8 @@
 #include "cuda_compat.h"
 #include "compiler.h"
 #include "print.h"
+#include "types.h"
 
-/**
- * Bfloat16 Arithmetic Constants
- * https://docs.nvidia.com/cuda/cuda-math-api/cuda_math_api/group__CUDA__MATH__INTRINSIC__BFLOAT16__CONSTANTS.html
- */
-#ifndef __NVCC__
-#define CUDART_INF_BF16	__ushort_as_bfloat16((unsigned short)0x7F80U)
-#define CUDART_MAX_NORMAL_BF16	__ushort_as_bfloat16((unsigned short)0x7F7FU)
-#define CUDART_MIN_DENORM_BF16	__ushort_as_bfloat16((unsigned short)0x0001U)
-#define CUDART_NAN_BF16	__ushort_as_bfloat16((unsigned short)0x7FFFU)
-#define CUDART_NEG_ZERO_BF16	__ushort_as_bfloat16((unsigned short)0x8000U)
-#define CUDART_ONE_BF16	__ushort_as_bfloat16((unsigned short)0x3F80U)
-#define CUDART_ZERO_BF16	__ushort_as_bfloat16((unsigned short)0x0000U)
-#endif
 
 __global__ void k_bfloat16_constants(void)
 {